trading-tools/requirements.txt
requests>=2.31
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
streamlit>=1.30
plotly>=5.18
matplotlib>=3.8
optuna>=3.5
torch>=2.1
xgboost>=2.0
python-dotenv>=1.0
pyyaml>=6.0
